// bot/src/types/openapi.ts
export interface components {
    schemas: {
// ...
        FoldResponse: {
            /** Turn */
            turn: {
                [key: string]: unknown;
            };
            /** Action */
            action: string;
            detail?: components["schemas"]["FoldDetail"] | null;
        };
        /** FormAnswerResponse */
        FormAnswerResponse: {
            /** Turn */
            turn: {
                [key: string]: unknown;
            };
            /**
             * Complete
             * @default false
             */
            complete: boolean;
            /** Answer */
            answer?: {
                [key: string]: unknown;
            };
            /** Missing */
            missing?: number[];
        };
// ...
        HtmlResponse: {
            /**
             * Html
             * @default
             */
            html: string;
        };
// ...
        NewTurnResponse: {
            /** Turn */
            turn: {
                [key: string]: unknown;
            };
        };
// ...
        RenderMarkdownResponse: {
            /**
             * Html
             * @default
             */
            html: string;
        };
// ...
        RenderRunningResponse: {
            /**
             * Text
             * @default
             */
            text: string;
            /** Keyboard */
            keyboard?: components["schemas"]["TurnButton"][][];
        };
// ...
        SafeFilenameResponse: {
            /**
             * Filename
             * @default
             */
            filename: string;
        };
// ...
        SelectOptionResponse: {
            /** Turn */
            turn: {
                [key: string]: unknown;
            };
            /**
             * Changed
             * @default false
             */
            changed: boolean;
        };
// ...
        SplitResponse: {
            /** Chunks */
            chunks?: string[];
        };
// ...
        TextResponse: {
            /**
             * Text
             * @default
             */
            text: string;
        };
        /** TurnButton */
        TurnButton: {
            /** Text */
            text: string;
            /** Data */
            data: string;
        };
        /** TurnPayload */
        TurnPayload: {
            /** Turn */
            turn: {
                [key: string]: unknown;
            };
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
            /** Input */
            input?: unknown;
            /** Context */
            ctx?: Record<string, never>;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}